リファクタ：バックエンドの処理を再構築
