+++ src/app_const.py
@@ -1,0 +1,12 @@
+# DalleCraft
+#
+# Copyright (c) 2025 led-mirage
+# このソースコードは MITライセンス の下でライセンスされています。
+# ライセンスの詳細については、このプロジェクトのLICENSEファイルを参照してください。
+
+APP_NAME = "DalleCraft"
+APP_VERSION = "1.0.1"
+COPYRIGHT = "Copyright 2025 led-mirage"
+
+IMAGE_DIR = "images"
+CONFIG_FILE = "config.json"
