+++ version.yaml
@@ -1,4 +1,4 @@
-Version: 1.0.0
+Version: 1.0.1
 CompanyName: led-mirage
 FileDescription: AI画像生成ツール（DALL-E3）
 InternalName: DalleCraft
